# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Blood Pressure Tracker Mobile API
  version: 0.1.0
  description: |
    API for mobile blood pressure tracking. Users can create an account, log in,
    upload tonometer photos for recognition, review pending/recognized/failed
    readings, save recognized readings, retrieve original images, and browse
    saved measurement history.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Auth
    description: Account creation and login
  - name: Measurements
    description: Measurement upload, review, image retrieval, save, and history
paths:
  /api/v1/signin:
    post:
      tags:
        - Auth
      summary: Create an account
      description: Creates a new user account with email and password and returns a bearer access token that expires according to the configured access-token TTL (default seven days).
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: correct horse battery staple
      responses:
        '201':
          description: Account created and bearer token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                default:
                  value:
                    accessToken: opaque-token-value
                    tokenType: Bearer
                    expiresAt: '2026-05-27T12:00:00.000Z'
                    user:
                      id: usr_123
                      email: user@example.com
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email is already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateEmail:
                  value:
                    error: conflict
                    message: Email is already registered
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/login:
    post:
      tags:
        - Auth
      summary: Log in
      description: Authenticates an existing user and returns a bearer access token that expires according to the configured access-token TTL (default seven days).
      operationId: logIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            examples:
              default:
                value:
                  email: user@example.com
                  password: correct horse battery staple
      responses:
        '201':
          description: User authenticated and bearer token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                default:
                  value:
                    accessToken: opaque-token-value
                    tokenType: Bearer
                    expiresAt: '2026-05-27T12:00:00.000Z'
                    user:
                      id: usr_123
                      email: user@example.com
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/measurements:
    post:
      tags:
        - Measurements
      summary: Upload a measurement image
      description: Stores an authenticated JPEG/PNG image, creates a pending measurement, and schedules recognition.
      operationId: uploadMeasurement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: JPEG or PNG image greater than 0 bytes and no more than 10 MB.
      responses:
        '201':
          description: Measurement created and queued for recognition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementUploadResponse'
              examples:
                default:
                  value:
                    id: msr_123
                    status: pending
                    measurementTime: '2026-05-27T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags:
        - Measurements
      summary: List measurement history
      description: Returns a paginated list of the authenticated user's recognized and saved measurements. Original image binary data is never included.
      operationId: listMeasurements
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Measurement history page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementHistoryPage'
              examples:
                default:
                  value:
                    items:
                      - id: msr_123
                        status: saved
                        systolic: 120
                        diastolic: 80
                        pulse: 68
                        armSide: left
                        measurementTime: '2026-05-27T12:00:00.000Z'
                        savedAt: '2026-05-27T12:05:00.000Z'
                      - id: msr_124
                        status: recognized
                        systolic: 121
                        diastolic: 81
                        pulse: 69
                        armSide: right
                        measurementTime: '2026-05-27T12:10:00.000Z'
                        savedAt: null
                    page: 1
                    pageSize: 20
                    hasNextPage: false
                    filters:
                      from: '2026-05-01T00:00:00.000Z'
                      to: '2026-05-31T23:59:59.999Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/measurements/{id}:
    get:
      tags:
        - Measurements
      summary: Get measurement detail
      description: Returns one owned measurement, including status, recognized values when available, and an original image link.
      operationId: getMeasurementDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeasurementId'
      responses:
        '200':
          description: Measurement detail for pending, recognized, failed, or saved state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementDetail'
              examples:
                pending:
                  value:
                    id: msr_123
                    status: pending
                    measurementTime: '2026-05-27T12:00:00.000Z'
                    imageUrl: /api/v1/measurements/msr_123/image
                recognized:
                  value:
                    id: msr_123
                    status: recognized
                    systolic: 120
                    diastolic: 80
                    pulse: 68
                    armSide: left
                    measurementTime: '2026-05-27T12:00:00.000Z'
                    imageUrl: /api/v1/measurements/msr_123/image
                failed:
                  value:
                    id: msr_123
                    status: failed
                    measurementTime: '2026-05-27T12:00:00.000Z'
                    recognitionError: Measurement could not be recognized from this image.
                    imageUrl: /api/v1/measurements/msr_123/image
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/measurements/{id}/save:
    post:
      tags:
        - Measurements
      summary: Save a recognized measurement
      description: Confirms a recognized owned measurement so it appears in default history.
      operationId: saveMeasurement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeasurementId'
      responses:
        '201':
          description: Measurement saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedMeasurement'
              examples:
                default:
                  value:
                    id: msr_123
                    status: saved
                    systolic: 120
                    diastolic: 80
                    pulse: 68
                    armSide: left
                    measurementTime: '2026-05-27T12:00:00.000Z'
                    savedAt: '2026-05-27T12:05:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Measurement is not in a recognized state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidState:
                  value:
                    error: conflict
                    message: Measurement must be recognized before it can be saved
  /api/v1/measurements/{id}/override:
    post:
      tags:
        - Measurements
      summary: Override recognized values
      description: Overrides one or more recognized reading fields on an owned recognized or saved measurement.
      operationId: overrideMeasurement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeasurementId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementOverrideRequest'
            examples:
              default:
                value:
                  systolic: 121
                  diastolic: 81
                  pulse: 69
      responses:
        '201':
          description: Measurement override persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverriddenMeasurement'
              examples:
                default:
                  value:
                    id: msr_123
                    status: recognized
                    systolic: 121
                    diastolic: 81
                    pulse: 69
                    armSide: right
                    measurementTime: '2026-05-27T12:00:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Measurement is not in a recognized/saved state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidState:
                  value:
                    error: conflict
                    message: Measurement must be recognized before override can be applied
  /api/v1/measurements/{id}/image:
    get:
      tags:
        - Measurements
      summary: Get original measurement image
      description: Returns the original JPEG/PNG image binary for an owned measurement.
      operationId: getMeasurementImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MeasurementId'
      responses:
        '200':
          description: Original image bytes.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token returned by signin or login, expiring according to ACCESS_TOKEN_TTL_SECONDS (default seven days).
  parameters:
    MeasurementId:
      name: id
      in: path
      required: true
      description: Measurement identifier.
      schema:
        type: string
        example: msr_123
    Page:
      name: page
      in: query
      required: false
      description: Positive page number. Defaults to 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Positive page size. Defaults to 20 and is capped by the implementation.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    From:
      name: from
      in: query
      required: false
      description: Inclusive ISO-8601 measurement time lower bound.
      schema:
        type: string
        format: date-time
        example: '2026-05-01T00:00:00.000Z'
    To:
      name: to
      in: query
      required: false
      description: Inclusive ISO-8601 measurement time upper bound.
      schema:
        type: string
        format: date-time
        example: '2026-05-31T23:59:59.999Z'
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error: validation_error
                message: Human-readable user-safe message
    Unauthorized:
      description: Missing, expired, revoked, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error: unauthorized
                message: Bearer token is invalid or expired
    NotFound:
      description: Resource does not exist or is not owned by the authenticated user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error: not_found
                message: Measurement was not found
    RateLimited:
      description: Too many authentication attempts for the client and email in the current rate-limit window.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                error: rate_limited
                message: Too many authentication attempts; try again later
  schemas:
    AuthRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: correct horse battery staple
    AuthResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - tokenType
        - expiresAt
        - user
      properties:
        accessToken:
          type: string
          example: opaque-token-value
        tokenType:
          type: string
          const: Bearer
        expiresAt:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - email
      properties:
        id:
          type: string
          example: usr_123
        email:
          type: string
          format: email
          example: user@example.com
    MeasurementUploadResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - measurementTime
      properties:
        id:
          type: string
          example: msr_123
        status:
          type: string
          const: pending
        measurementTime:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
    MeasurementStatus:
      type: string
      enum:
        - pending
        - recognizing
        - recognized
        - saved
        - failed
    ArmSide:
      type: string
      enum:
        - left
        - right
        - unknown
    MeasurementDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - measurementTime
        - imageUrl
      properties:
        id:
          type: string
          example: msr_123
        status:
          $ref: '#/components/schemas/MeasurementStatus'
        systolic:
          type: integer
          example: 120
        diastolic:
          type: integer
          example: 80
        pulse:
          type: integer
          example: 68
        armSide:
          $ref: '#/components/schemas/ArmSide'
        measurementTime:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
        recognitionError:
          type: string
          example: Measurement could not be recognized from this image.
        imageUrl:
          type: string
          example: /api/v1/measurements/msr_123/image
        savedAt:
          type: string
          format: date-time
          example: '2026-05-27T12:05:00.000Z'
    SavedMeasurement:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - systolic
        - diastolic
        - pulse
        - armSide
        - measurementTime
        - savedAt
      properties:
        id:
          type: string
          example: msr_123
        status:
          type: string
          const: saved
        systolic:
          type: integer
          example: 120
        diastolic:
          type: integer
          example: 80
        pulse:
          type: integer
          example: 68
        armSide:
          $ref: '#/components/schemas/ArmSide'
        measurementTime:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
        savedAt:
          type: string
          format: date-time
          example: '2026-05-27T12:05:00.000Z'
    MeasurementOverrideRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        systolic:
          type: integer
          minimum: 1
          example: 121
        diastolic:
          type: integer
          minimum: 1
          example: 81
        pulse:
          type: integer
          minimum: 1
          example: 69
    OverriddenMeasurement:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - systolic
        - diastolic
        - pulse
        - armSide
        - measurementTime
      properties:
        id:
          type: string
          example: msr_123
        status:
          type: string
          enum:
            - recognized
            - saved
        systolic:
          type: integer
          example: 121
        diastolic:
          type: integer
          example: 81
        pulse:
          type: integer
          example: 69
        armSide:
          $ref: '#/components/schemas/ArmSide'
        measurementTime:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
        savedAt:
          type: string
          format: date-time
          example: '2026-05-27T12:05:00.000Z'
    MeasurementHistoryItem:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - systolic
        - diastolic
        - pulse
        - armSide
        - measurementTime
        - savedAt
      properties:
        id:
          type: string
          example: msr_123
        status:
          type: string
          enum:
            - recognized
            - saved
        systolic:
          type: integer
          example: 120
        diastolic:
          type: integer
          example: 80
        pulse:
          type: integer
          example: 68
        armSide:
          $ref: '#/components/schemas/ArmSide'
        measurementTime:
          type: string
          format: date-time
          example: '2026-05-27T12:00:00.000Z'
        savedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-27T12:05:00.000Z'
    MeasurementHistoryPage:
      type: object
      additionalProperties: false
      required:
        - items
        - page
        - pageSize
        - hasNextPage
        - filters
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementHistoryItem'
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          example: 20
        hasNextPage:
          type: boolean
          example: false
        filters:
          type: object
          additionalProperties: false
          required:
            - from
            - to
          properties:
            from:
              type:
                - string
                - 'null'
              format: date-time
              example: '2026-05-01T00:00:00.000Z'
            to:
              type:
                - string
                - 'null'
              format: date-time
              example: '2026-05-31T23:59:59.999Z'
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - validation_error
            - unauthorized
            - not_found
            - conflict
            - rate_limited
        message:
          type: string
          example: Human-readable user-safe message